bug 14362: VK_NULL_HANDLE does not work in C (WIP)
Drop type safety on 32-bit.

https://cvs.khronos.org/bugzilla/show_bug.cgi?id=14362

+++ icd/intel/cmd.c
@@ -501,7 +501,7 @@ void intel_free_cmd_buffers(
     const VkCmdBuffer                  *cmd_bufs)
 {
     for (uint32_t i = 0; i < count; i++) {
-        struct intel_obj *obj = intel_obj((uint64_t)cmd_bufs[i]);
+        struct intel_obj *obj = intel_obj(cmd_bufs[i]);
 
         obj->destroy(obj);
     }
